# remove support for dynamic choice of B in BTree, significantly reducing ...