@huonw 2015-02-12
Squishing the pointers doesn't seem to be worth it, IMO, that is, it's a fairly significant performance suck for the iteration.