@Gankro 2015-02-12
The B=6 test shows the same perf as the original. Variance is pretty high on those benches, though. All the results are within a std deviation.