@Gankro 2015-02-15
This got super clobbered by phantom stuff, and I want to investigate using arrays more now. Closing.